requirements.txt
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.0
click==8.1.7
cryptography==41.0.5
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
docutils==0.20.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.34.4
fsspec==2023.10.0
google-api-core==2.14.0
google-auth==2.23.4
google-auth-oauthlib==0.5.2
google-cloud-core==2.3.3
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
gspread==5.4.0
gunicorn==21.2.0
idna==3.3
importlib-metadata==7.0.1
inflection==0.5.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
keyring==24.3.0
kiwisolver==1.4.4
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
more-itertools==8.14.0
mpmath==1.3.0
multitasking==0.0.11
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.2.1
nh3==0.2.15
nltk==3.8.1
numpy==1.23.1
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
pdfquery==0.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipdeptree==2.13.1
pkginfo==1.9.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.22.3
protobuf==4.25.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.0.9
pyquery==2.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0.1
pyzmq==23.2.0
readme-renderer==42.0
regex==2023.8.8
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
roman==4.1
rsa==4.9
safetensors==0.4.0
scipy==1.9.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
sympy==1.12
terminado==0.15.0
threadpoolctl==3.2.0
tinycss2==1.1.1
tornado==6.2
toto-pdfquery==0.1.3
tqdm==4.66.1
traitlets==5.3.0
twine==4.0.2
typing_extensions==4.8.0
Unidecode==1.3.4
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.3.7
zipp==3.17.0
totoapicontroller==1.0.0